+++ stdlib/Dates/test/types.jl
@@ -235,7 +235,8 @@ end
     for (a, b) in [(Dates.Date(2000), Dates.Date(2001)),
                     (Dates.Time(10), Dates.Time(11)),
                     (Dates.DateTime(3000), Dates.DateTime(3001)),
-                    (Dates.Week(42), Dates.Week(1972))]
+                    (Dates.Week(42), Dates.Week(1972)),
+                    (Dates.Quarter(3), Dates.Quarter(52))]
         @test min(a, b) == a
         @test min(b, a) == a
         @test min(a) == a